Refactor and add option to get geometries from a set of shapefile in a directory
